test(rs-platform-wallet/e2e): invert ID-007 — assert correct behavior, fail until upstream fixes BlockchainIdentities
ID-007 previously pinned the broken contract — green-while-buggy. That
inverts the meaning of the test suite: a green run could mean either
"feature works" or "feature still broken in the same way as before",
and there's no way to tell at a glance.

Flip the polarity. The test now asserts the CORRECT behavior:

- identity-auth addresses ARE in `monitored_addresses()` before and
  after the Layer-1 send,
- the wallet's confirmed Core balance INCREASES after the inbound
  UTXO confirms,
- the wallet's UTXO set CONTAINS the new entry.

All three assertions currently FAIL because rust-dashcore's
`WalletAccountCreationOptions::Default` excludes
`BlockchainIdentities*` `AccountType` variants at the pinned
`key-wallet` revision (closed PR `dashpay/rust-dashcore#554` attempted
this; closed without merge). The negative-axis variant
(`identity_index = 1`, unregistered slot) carries the same
correct-behavior assertions — registration status is irrelevant to
monitoring since the derivation is pure.

Renames the test file and function from `..._not_monitored` to
`..._monitored` to match the inverted intent. Bumps
`wait_for_core_balance` timeout to 5 minutes (testnet block time
~2.5 min plus SPV bloom-filter propagation headroom) since
the assertion is now "balance reaches target", not "wait times out".
The `#[ignore]` reason now spells out "FAILS by design until
upstream lands BlockchainIdentities* support".

DET parallel: `dash-evo-tool#692`.
